+++ CHANGES.rst
@@ -1,6 +1,15 @@
 Changes
 =======
 
+0.3.0 -- 2019-mm-dd
+-------------------
+* Performance improvements (evaluation of mask scores)
+* Faster PNG output
+* Deprecation of ``encoder.score_n1``, ``encoder.score_n2``, ``encoder.score_n3``,
+  and ``encoder.score_n4``.
+  Use ``encoder.mask_scores`` or ``encoder.evaluate_mask``.
+
+
 0.2.9 -- 2019-04-24
 -------------------
 * Fixed typos